requirements.txt
adal==1.2.3
azure-common==1.1.4
azure-mgmt-compute==12.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==0.31.0
azure-mgmt-sql==0.2.0
azure-mgmt-subscription==0.5.0
azure-nspkg==3.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
haikunator==2.1.0
idna==2.9
isodate==0.6.0
keyring==21.2.1
msrest==0.6.13
msrestazure==0.4.34
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
termcolor==1.1.0
urllib3==1.25.9
